Added SAVE functionality for scores
Also modified the back-end controllers with new functionality.

+++ src/org/sw/marketing/servlet/SelfAssessmentController.java
@@ -16,6 +16,7 @@ import org.sw.marketing.dao.DAOFactory;
 import org.sw.marketing.dao.answer.AnswerDAO;
 import org.sw.marketing.dao.form.FormDAO;
 import org.sw.marketing.dao.question.QuestionDAO;
+import org.sw.marketing.dao.score.ScoreDAO;
 import org.sw.marketing.dao.submission.SubmissionAnswerDAO;
 import org.sw.marketing.dao.submission.SubmissionDAO;
 import org.sw.marketing.dao.user.UserDAO;
@@ -25,9 +26,11 @@ import org.sw.marketing.data.form.Data.Form;
 import org.sw.marketing.data.form.Data.Form.Question;
 import org.sw.marketing.data.form.Data.Form.Question.PossibleAnswer;
 import org.sw.marketing.data.form.Data.Message;
+import org.sw.marketing.data.form.Data.Score;
 import org.sw.marketing.data.form.Data.Submission;
 import org.sw.marketing.data.form.Data.Submission.Answer;
 import org.sw.marketing.servlet.params.QuestionParameters;
+import org.sw.marketing.servlet.params.ScoreParameters;
 import org.sw.marketing.servlet.params.SurveyParameters;
 import org.sw.marketing.transformation.TransformerHelper;
 import org.sw.marketing.util.ReadFile;
@@ -46,6 +49,7 @@ public class SelfAssessmentController extends HttpServlet
 		innerScreenList.add("QUESTIONS_AND_ANSWERS");
 		innerScreenList.add("ANSWERS");
 		innerScreenList.add("SCORES");
+		innerScreenList.add("EDIT_SCORE");
 		innerScreenList.add("REPORTS");
 		innerScreenList.add("ANALYTICS");
 	}
@@ -61,6 +65,8 @@ public class SelfAssessmentController extends HttpServlet
 		FormDAO formDAO = DAOFactory.getFormDAO();
 		QuestionDAO questionDAO = DAOFactory.getQuestionDAO();
 		AnswerDAO answerDAO = DAOFactory.getPossibleAnswerDAO();
+		ScoreDAO scoreDAO = DAOFactory.getScoreDAO();
+		SubmissionDAO submissionDAO = DAOFactory.getSubmissionDAO();
 		
 		/*
 		 * Data Initialization
@@ -68,9 +74,11 @@ public class SelfAssessmentController extends HttpServlet
 		Data data = new Data();
 		java.util.List<Form> formList = null;
 		Form form = null;
-		java.util.List<Question> questionList = null;
+		java.util.List<Question> questionList = null;		
 		Question question = null;
-		java.util.List<PossibleAnswer> possibleAnswerList = null;
+		java.util.List<Data.PossibleAnswer> possibleAnswers = null;				
+		java.util.List<Score> scores = null;
+		Score score = null;
 		User user = null;
 		
 		/*
@@ -111,6 +119,10 @@ public class SelfAssessmentController extends HttpServlet
 			formID = (long) request.getAttribute("surveyId");
 			request.removeAttribute("surveyId");
 		}
+		if(formID > 0)
+		{
+			form = formDAO.getForm(formID);
+		}
 		
 		/*
 		 * Question ID
@@ -128,6 +140,28 @@ public class SelfAssessmentController extends HttpServlet
 			{
 				questionID = null;
 			}
+		}
+		
+		/*
+		 * Score ID
+		 */
+		String scoreIdStr = null;
+		long scoreID = 0;
+		if(parameterMap.get("SCORE_ID") != null)
+		{
+			scoreIdStr = parameterMap.get("SCORE_ID")[0];
+			try
+			{
+				scoreID = Long.parseLong(scoreIdStr);
+			}
+			catch(NumberFormatException e)
+			{
+				scoreID = 0;
+			}
+		}
+		if(scoreID > 0)
+		{
+			score = scoreDAO.getScore(scoreID);
 		}
 		
 		/*
@@ -157,26 +191,9 @@ public class SelfAssessmentController extends HttpServlet
 				formID = 0;
 			}
 			else if(paramAction.equals("SAVE_FORM"))
-			{
-				form = SurveyParameters.process(request, formDAO.getForm(formID));
-				
-				Form tempForm = formDAO.getFormByPrettyUrl(form.getPrettyUrl());
-				if(tempForm != null && tempForm.getId() != form.getId() && tempForm.getPrettyUrl().equals(form.getPrettyUrl()))
-				{
-					Message message = new Message();
-					message.setType("error");
-					message.setLabel("The pretty URL is already in use.  Please choose a unique one.");
-					data.getMessage().add(message);
-				}
-				else
-				{
-					formDAO.updateForm(form);
-					
-					Message message = new Message();
-					message.setType("success");
-					message.setLabel("The form has been saved.");
-					data.getMessage().add(message);
-				}
+			{				
+				form = SurveyParameters.process(request, form);
+				formDAO.updateForm(form);
 			}
 			else if(paramAction.equals("CREATE_QUESTION"))
 			{
@@ -254,19 +271,30 @@ public class SelfAssessmentController extends HttpServlet
 			}
 			else if(paramAction.equals("DELETE_QUESTION"))
 			{
-				question = questionDAO.getQuestion(questionID);
-				int count = questionDAO.getQuestionPageCount(question.getPage());
-				if(count == 1)
+				java.util.List<Submission> submissions = submissionDAO.getSubmissions(formID);
+				if(submissions != null)
 				{
-					questionDAO.removePageBreak(question.getNumber(), formID);
+					Message message = new Message();
+					message.setType("error");
+					message.setLabel("You cannot delete questions that have submissions. Consider creating a new self-assessment.");
+					data.getMessage().add(message);
 				}
-				questionDAO.deleteQuestion(questionID);
-				questionDAO.moveUpQuestions(question.getNumber(), formID);
-				
-				Message message = new Message();
-				message.setType("error");
-				message.setLabel("The question has been deleted.");
-				data.getMessage().add(message);
+				else
+				{
+					question = questionDAO.getQuestion(questionID);
+					int count = questionDAO.getQuestionPageCount(question.getPage());
+					if(count == 1)
+					{
+						questionDAO.removePageBreak(question.getNumber(), formID);
+					}
+					questionDAO.deleteQuestion(questionID);
+					questionDAO.moveUpQuestions(question.getNumber(), formID);
+					
+					Message message = new Message();
+					message.setType("success");
+					message.setLabel("The question has been deleted.");
+					data.getMessage().add(message);
+				}
 			}
 			else if(paramAction.equals("INSERT_PAGE_BREAK"))
 			{
@@ -367,6 +395,33 @@ public class SelfAssessmentController extends HttpServlet
 				long answerID = Long.parseLong(parameterMap.get("ANSWER_ID")[0]);
 				answerDAO.deleteAnswerForForm(answerID);
 			}
+			else if(paramAction.equals("CREATE_SCORE"))
+			{
+				scoreID = scoreDAO.insertScore(formID);
+				score = scoreDAO.getScore(scoreID);
+			}
+			else if(paramAction.equals("SAVE_SCORE"))
+			{
+				score = ScoreParameters.process(request, score);
+				scoreDAO.updateScore(score);
+				
+				String beginScoreStr = parameterMap.get("SCORE_BEGIN")[0];
+				int beginScore = Integer.parseInt(beginScoreStr);
+				String endScoreStr = parameterMap.get("SCORE_END")[0];
+				int endScore = Integer.parseInt(endScoreStr);
+				
+				if(beginScore > endScore)
+				{
+					Message message = new Message();
+					message.setType("error");
+					message.setLabel("The low score must be smaller than or equal to the high score.  The values have been swapped.");
+					data.getMessage().add(message);
+				}
+			}
+			else if(paramAction.equals("DELETE_SCORE"))
+			{
+				scoreDAO.deleteScore(scoreID);
+			}
 		}
 		
 		/*
@@ -387,37 +442,30 @@ public class SelfAssessmentController extends HttpServlet
 			}
 			else if(paramScreen.equals("QUESTIONS_AND_ANSWERS"))
 			{
-				questionList = questionDAO.getQuestions(formID);
 				xslScreen = "questions_and_answers.xsl";
-				if(questionList != null)
-				{
-					form.getQuestion().addAll(questionList);
-				}
 			}			
 			else if(paramScreen.equals("ANSWERS"))
 			{
-				java.util.List<Data.PossibleAnswer> possibleAnswers = answerDAO.getPossibleAnswersByForm(formID);
-				if(possibleAnswers != null)
-				{
-					data.getPossibleAnswer().addAll(possibleAnswers);
-				}
 				xslScreen = "answers.xsl";
 			}		
 			else if(paramScreen.equals("SCORES"))
-			{
-				questionList = questionDAO.getQuestions(formID);
-				xslScreen = "questions_and_answers.xsl";
-				if(questionList != null)
+			{				
+				scores = scoreDAO.getScores(formID);
+				if(scores != null)					
 				{
-					form.getQuestion().addAll(questionList);
-				}
-				
-				java.util.List<Data.PossibleAnswer> possibleAnswers = answerDAO.getPossibleAnswersByForm(formID);
-				if(possibleAnswers != null)
-				{
-					data.getPossibleAnswer().addAll(possibleAnswers);
+					data.getScore().addAll(scores);
 				}
 				xslScreen = "scores.xsl";
+			}	
+			else if(paramScreen.equals("EDIT_SCORE"))
+			{	
+				score = scoreDAO.getScore(scoreID);
+				if(score != null)					
+				{
+					data.getScore().add(score);
+				}			
+				
+				xslScreen = "score_edit.xsl";
 			}
 			else if(paramScreen.equals("REPORTS"))
 			{
@@ -434,6 +482,18 @@ public class SelfAssessmentController extends HttpServlet
 			
 			if(form != null)
 			{
+				questionList = questionDAO.getQuestions(formID);
+				if(questionList != null)
+				{
+					form.getQuestion().addAll(questionList);
+				}
+				
+				possibleAnswers = answerDAO.getPossibleAnswersByForm(formID);
+				if(possibleAnswers != null)
+				{
+					data.getPossibleAnswer().addAll(possibleAnswers);
+				}
+				
 				data.getForm().add(form);
 			}
 		}
